falta desenhar

+++ CaminhosMarte/app/src/main/java/com/example/caminhosmarte/DistOriginal.java
@@ -1,7 +1,7 @@
 package com.example.caminhosmarte;
 
 public class DistOriginal {
-    public int distancia;
+    public long distancia;
     public int verticePai;
 
     public DistOriginal(int vp, int d)
